Wraparound fixed
Fixed wraparound issues. Now wraparound is smooth.

When going North or East, the wraparound check needs to be > instead of >= because otherwise it'll simulate a move, see you on the edge, then wrap you around.

For the actual location, instead of 0, TilePixelHeight and width are used. Using 0 will cause indexing issues, and makes the character look like he's hanging onto the edge of the map, on the outside.

Whenwrapping south and west, same deal. The check uses <=0, because hitting 0 will put you out of bounds. Destinatoon changed to be on the edge instead of 1 row behind it.